+++ App/System/Clock.cpp
@@ -11,12 +11,26 @@ Clock::Clock(ClockTimer &timer) :
     resetTicks();
     _ppqn = CONFIG_PPQN;
     _masterBpm = 120.f;
+    _mode = Mode::Auto;
     _runState = RunState::Idle;
+    _slaveDivisor = CONFIG_PPQN / 24; // -> 8
     _timer.attach(this);
 }
 
 void Clock::init() {
     _timer.disable();
+}
+
+void Clock::setMode(Mode mode) {
+    if(mode != _mode) {
+        if(mode == Mode::Master && _runState == RunState::SlaveRunning) {
+            slaveStop();
+        }
+        if(mode == Mode::Slave && _runState == RunState::MasterRunning) {
+            masterStop();
+        }
+        _mode = mode;
+    }
 }
 
 void Clock::masterStart() {
@@ -36,32 +50,114 @@ void Clock::masterStop() {
 void Clock::resetTicks() {
     _tick = 0;
     _tickProcessed = 0;
+    _slaveSubTicksPending = 0;
     _output.nextTick = 0;
 }
 
 void Clock::setMasterBpm(float bpm) {
     _masterBpm = bpm;
-    setupMasterTimer();
+    if(_runState == RunState::MasterRunning) {
+        setupMasterTimer();
+    }
 }
 
+void Clock::slaveTick() {
+    if(_runState == RunState::SlaveRunning) {
+        // protect against clock rate overload
+        _slaveSubTicksPending = std::min(_slaveSubTicksPending + _slaveDivisor, 2 * _slaveDivisor);
+        
+        // time elapsed since last tick
+        uint32_t periodUs = _elapsedUs - _lastSlaveTickUs;
+
+        // default tick period to 120 bpm
+        if(_slaveTickPeriodUs == 0) {
+            _slaveTickPeriodUs = (60 * 1000000 * _slaveDivisor) / (120 * _ppqn);
+        }
+
+        // update tick period if we have a valid measurement
+        if(periodUs > 0 && _lastSlaveTickUs > 0) {
+            _slaveTickPeriodUs = periodUs;
+        }
+
+        _slaveSubTickPeriodUs = _slaveTickPeriodUs / _slaveSubTicksPending;
+        if(_elapsedUs - _nextSlaveSubTickUs > 1000) {
+            _nextSlaveSubTickUs = _elapsedUs;
+        } else {
+            _nextSlaveSubTickUs += _slaveSubTickPeriodUs;
+        }
+
+        // estimate slave bpm
+        if(periodUs > 0 && _lastSlaveTickUs > 0) {
+            float bpm = (60.f * 1000000 * _slaveDivisor) / (periodUs * _ppqn);
+            _slaveBpmFiltered = .9f * _slaveBpmFiltered + .1f * bpm;
+            _slaveBpmAvg.push(_slaveBpmFiltered);
+            _slaveBpm = _slaveBpmAvg();
+            //USBDBG("E:%ld, L:%ld, B:%.2f, S: %.2f\n", _elapsedUs, _lastSlaveTickUs, bpm, _slaveBpm);
+        }
+
+        _lastSlaveTickUs = _elapsedUs;
+    }
+}
+
+void Clock::slaveStart() {
+    if(_runState == RunState::MasterRunning || _mode == Mode::Master) {
+        return;
+    }
+
+    setRunState(RunState::SlaveRunning);
+
+    resetTicks();
+
+    _timer.disable();
+    setupSlaveTimer();
+    _timer.enable();
+}
+
+void Clock::slaveStop() {
+    if(_runState != RunState::SlaveRunning || _mode == Mode::Master) {
+        return;
+    }
+
+    setRunState(RunState::Idle);
+
+    _timer.disable();
+}
+
+void Clock::slaveContinue() {
+    if(_runState != RunState::Idle || _mode == Mode::Master) {
+        return;
+    }
+
+    setRunState(RunState::SlaveRunning);
+
+    setupSlaveTimer();
+    _timer.enable();
+}
+
+void Clock::slaveReset() {
+    if(_runState == RunState::MasterRunning || _mode == Mode::Master) {
+        return;
+    }
+
+    setRunState(RunState::Idle);
+
+    _timer.disable();
+}
+
+// called by USB ISR
 void Clock::slaveHandleMidi(uint8_t msg) {
     switch (MidiMessage::realTimeMessage(msg)) {
     case MidiMessage::Tick:
-        //slaveTick(slave);
+        slaveTick();
         break;
     case MidiMessage::Start:
-        USBDBG("MIDI Start\n");
-        //slaveStart(slave);
-        masterStart();
+        slaveStart();
         break;
     case MidiMessage::Stop:
-        USBDBG("MIDI Stop\n");
-        //slaveStop(slave);
-        masterStop();
+        slaveStop();
         break;
     case MidiMessage::Continue:
-        USBDBG("MIDI Continue\n");
-        //slaveContinue(slave);
+        slaveContinue();
         break;
     default:
         break;
@@ -130,6 +226,22 @@ void Clock::onClockTimerTick() {
         _elapsedUs += _timer.period();
         break;
     }
+    case RunState::SlaveRunning: {
+        _elapsedUs += _timer.period();
+
+        if(_slaveSubTicksPending > 0 && _elapsedUs >= _nextSlaveSubTickUs) {
+            outputTick(_tick);
+            ++_tick;
+            --_slaveSubTicksPending;
+            _nextSlaveSubTickUs += _slaveSubTickPeriodUs;
+        }
+
+        if(_mode == Mode::Auto && (_elapsedUs - _lastSlaveTickUs) > 500000) {
+            USBDBG("Auto Slave Reset\n");
+            slaveReset();
+        }
+        break;
+    }
     default:
         break;
     }
@@ -141,6 +253,13 @@ void Clock::setupMasterTimer() {
     // From 181 onwards the difference in period can be zero from one tenth bpm to the next.
     uint32_t us = std::round((60 * 1000000) / (_masterBpm * _ppqn));
     _timer.setPeriod(us);
+}
+
+void Clock::setupSlaveTimer() {
+    _elapsedUs = 0;
+    _lastSlaveTickUs = 0;
+
+    _timer.setPeriod(SlaveTimerPeriod); // 100 -> 3125 bpm ?
 }
 
 // called from onClockTimerTick