+++ App/System/Clock.h
@@ -2,6 +2,7 @@
 
 #include "ClockTimer.h"
 #include "Config.h"
+#include "MovingAverage.h"
 #include <list>
 
 class IClockObserver {
@@ -26,22 +27,46 @@ public:
 
 class Clock : public IClockTimerObserver, public IClock  {
 public:
+    enum class Mode {
+        Auto,
+        Master,
+        Slave
+    };
+
+    enum class RunState {
+        Idle,
+        MasterRunning,
+        SlaveRunning
+    };
 
     Clock(ClockTimer &timer);
 
     void init();
 
+    Mode mode() const { return _mode; }
+    void setMode(Mode mode);
+
+    Mode activeMode() const;
+
+    inline RunState runState() const { return _runState; }
     inline bool isIdle() const { return _runState == RunState::Idle; }
     inline bool isRunning() const { return _runState != RunState::Idle; }
     inline uint32_t ppqn() { return _ppqn; }
-    inline uint32_t bpm() const { return masterBpm(); }
+    inline uint32_t bpm() const { return _runState == RunState::SlaveRunning ? slaveBpm() : masterBpm(); }
 
     // Master clock control
     void masterStart();
     void masterStop();
     inline float masterBpm() const { return _masterBpm; }
     void setMasterBpm(float bpm);
+    inline float slaveBpm() const { return _slaveBpm; }
 
+    // Slave clock control
+    void slaveTick();
+    void slaveStart();
+    void slaveStop();
+    void slaveContinue();
+    void slaveReset();
     void slaveHandleMidi(uint8_t msg);
 
     uint32_t tick() const { return _tick; }
@@ -67,11 +92,6 @@ public:
     void onClockTimerTick() override;
 
 private:
-    enum class RunState {
-        Idle,
-        MasterRunning
-    };
-
     struct Output {
         int divisor;
         int pulse;
@@ -91,20 +111,38 @@ private:
     inline void setRunState(RunState runState) { _runState = runState; }
 
     void setupMasterTimer();
+    void setupSlaveTimer();
 
     void outputTick(uint32_t tick);
     void outputClock(bool clock);
 
+    static constexpr uint32_t SlaveTimerPeriod = 100; // us -> 3125 bpm
+
     ClockTimer &_timer;
+    // TODO: we can remove the concept of mode setting
+    Mode _mode;
     RunState _runState;
     IClockObserver::OutputState _outputState;
 
     volatile uint32_t _tick;
     volatile uint32_t _tickProcessed;
 
+    uint32_t _slaveDivisor;
+
     uint32_t _elapsedUs;
     float _masterBpm;
     uint32_t _ppqn;
+
+    // slave sync helpers
+    uint32_t _lastSlaveTickUs; // time of last call to slaveTick
+    uint32_t _slaveTickPeriodUs = 0; // slave tick period time
+    uint32_t _slaveSubTicksPending; // number of slave sub ticks pending
+    uint32_t _slaveSubTickPeriodUs = 0; // slave sub tick period time
+    uint32_t _nextSlaveSubTickUs; // time of next slave sub tick
+
+    float _slaveBpmFiltered = 0.f;
+    MovingAverage<float, 4> _slaveBpmAvg;
+    float _slaveBpm = 0.f;
 
     std::list<IClockObserver*> _observerList;
 };