+++ Middlewares/ST/STM32_USB_Device_Library/Core/Inc/usbd_def.h
@@ -189,6 +189,7 @@ extern "C" {
   * @}
   */
 
+
 /** @defgroup USBD_DEF_Exported_TypesDefinitions
   * @{
   */