+++ USB_DEVICE/App/usb_device.c
@@ -23,6 +23,7 @@
 #include "usb_device.h"
 #include "usbd_core.h"
 #include "usbd_desc.h"
+#include "usbd_cdc.h"
 #include "usbd_cdc_if.h"
 #include "usbd_midi_if.h"
 
@@ -69,7 +70,6 @@ void MX_USB_DEVICE_Init(void)
   /* USER CODE END USB_DEVICE_Init_PreTreatment */
 
   /* Init Device Library, add supported class and start the library. */
-
   if (USBD_Init(&hUsbDeviceFS, &FS_Desc, DEVICE_FS) != USBD_OK)
   {
     Error_Handler();