+++ USB_DEVICE/App/usbd_desc.h
@@ -46,8 +46,8 @@
   * @{
   */
 #define         DEVICE_ID1          (UID_BASE)
-#define         DEVICE_ID2          (UID_BASE + 0x4U)
-#define         DEVICE_ID3          (UID_BASE + 0x8U)
+#define         DEVICE_ID2          (UID_BASE + 0x4)
+#define         DEVICE_ID3          (UID_BASE + 0x8)
 
 #define  USB_SIZ_STRING_SERIAL       0x1A