+++ USB_DEVICE/Target/usbd_conf.h
@@ -80,7 +80,7 @@
   */
 
 /*---------- -----------*/
-#define USBD_MAX_NUM_INTERFACES     4U
+#define USBD_MAX_NUM_INTERFACES     3U
 /*---------- -----------*/
 #define USBD_MAX_NUM_CONFIGURATION     1U
 /*---------- -----------*/
@@ -93,12 +93,6 @@
 #define USBD_LPM_ENABLED     0U
 /*---------- -----------*/
 #define USBD_SELF_POWERED     1U
-/*---------- -----------*/
-#define USBD_DFU_MAX_ITF_NUM     1U
-/*---------- -----------*/
-#define USBD_DFU_XFER_SIZE     1024U
-/*---------- -----------*/
-#define USBD_DFU_APP_DEFAULT_ADD     0x08000000U
 
 /****************************************/
 /* #define for FS and HS identification */