+++ CHANGELOG.md
@@ -1,2 +1,4 @@
 v0.2.3
+- Add USB MIDI sync
 - Add USB composite interface (CDC + MIDI)
+- Add USB debug routine via CDC